Checking that basic php networking works
Tested a few things using the old php networking library, so far is promising. there are some things that obviously need changing (ie the fact that this is formerly a blog framework), plus the metadata element needs reworking. SQL db was also populated, made some scripts that update the file and the local db, depending on the need.